feat(transforms): add new transforms module
The transform module is a collection of differentiable transformations
that can be used to constrain an LFA to a specific range of values. For
example, one may want to squash the output values to the positive real
line via an exponential function. These all implement a `grad` method
which is used to correct updates as required.